experiment with OQSBuiler being a single script
