+++ requirements.txt
@@ -1,3 +1,5 @@
 PyYAML==6.0.2
 gersemi==0.23.1
 Jinja2==3.1.6
+pytest
+pytest-xdist
