+++ scripts/oqsbuilder_templates/kem_cmakelists.txt.jinja
