+++ scripts/oqsbuilder_templates/kem_src.c.jinja
