+++ src/kem/ml_kem/CMakeLists.txt
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MIT
-# This file was generated by oqsbuilder.oqsbuilder.KemFamily.generate_cmake
+# This file was generated by scripts/oqsbuilder.py:KemFamily.generate_cmake
 
 set(_ML_KEM_OBJS "")
