+++ src/kem/ml_kem/kem_ml_kem.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// This file was generated by oqsbuilder.oqsbuilder.KemFamily.generate_header
+// This file was generated by scripts/oqsbuilder.py:KemFamily.generate_header
 
 #ifndef OQS_KEM_ML_KEM_H
 #define OQS_KEM_ML_KEM_H
