+++ src/kem/ml_kem/kem_ml_kem_1024.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// This file was generated by oqsbuilder.oqsbuilder.KemFamily.generate_sources
+// This file was generated by scripts/oqsbuilder.py:KemFamily.generate_sources
 
 #include <stdlib.h>
